Rework a lot of skylark type handling in bazel_to_cmake.
This introduces types to for RepositoryId, PackageId, and TargetId,
which are used as keys in maps (which used to be strings), and as
anchor points to parse package and repository relative targets
(used for Label). In the process solve relative parsing of packages
and targets in some edge cases (the specific case that I saw was
involved package-relative label resolution in a library macro which
eventually was used in a select clause).

After that, decompose a lot of evaluation.py and related skylark types
into the skylark/ directory, adding tests. This renames the old
EvaluationContext class to EvaluationState, and introduces an abstract InvocationContext and a concrete EvaluationContext (forwarding to the
old implementation) to separate concerns and dependencies.

The basic evaluation procedure is essentially the same, however
when a deferred rule is created, the rule takes a context.snapshot()
(which copies the current state) rather than multi step package/package id
copies.

PiperOrigin-RevId: 491768918
Change-Id: Ic6fc92d0b46c08759f69c53ac188d45b35239bd5