Fix build errors on Python 3.11 due to PEP-670
On Python 3.11, tensorstore fails to build with the following error:

```
python/tensorstore/bfloat16.cc:774:31: error: expression is not assignable
    Py_TYPE(&NPyBfloat16_Descr) = &PyArrayDescr_Type;
    ~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^
  1 error generated.
```

As per [PEP-670], `Py_TYPE` is no longer a macro since python 3.11,
but now is a static inline function; we should instead use `Py_SET_TYPE`
to avoid the compilation error.

With this commit tensorstore can be built from source on Python 3.11.

[PEP-670]: https://peps.python.org/pep-0670/