Fix AsyncCache::TransactionNode initialization race
Previously, there was a race condition between
`AsyncCache::DoRequestWriteback` and the initialization of implicit
transaction nodes (`AsyncCache::implicit_transaction_node_`) that
could result in a null pointer dereference in
`AsyncCache::DoRequestWriteback`.

This commit fixes that race by ensuring that `DoRequestWriteback`
ignores a not-yet-initialized implicit transaction node.

Note that this does not create a risk of a writeback request from the
cache pool being lost, because if the implicit transaction is not yet
initialized then by definition there is not yet anything to write
back, and the request must have been stale (which can occur due to
releasing locks).

PiperOrigin-RevId: 506045735
Change-Id: I08e857fe6fb154fa0783bbf4b1f2800ce1b02453