OCDBT: Add support for numbered manifests in addition to single-file ones
The existing manifest format is only safe for concurrent writes on
kvstores that support atomic read-modify-write of single keys.

This commit adds a new manifest format that is safe for kvstores that
don't support general atomic read-modify-write of single keys, but do
support it in the special case where `if_equal=NoValue`.

PiperOrigin-RevId: 525321225
Change-Id: Ie52d48f113b7cae981a43386424175e99ce623c3