Change AsyncArray/ChunkCache to use SharedArray in place of SharedArrayView
Previously, the current state of a chunk was stored as a
`SharedArrayView` rather than a `SharedArray`.  The user of
`ChunkCache` was required to arrange for the storage of the chunk
`StridedLayout` separately.  This introduced some added complexity but
avoided redundantly storing the StridedLayout for the existing formats
where all chunks use the same StridedLayout for every chunk.  However,
for zarr v3, the restriction of using the same `StridedLayout` for
every chunk is awkward to implement and future codecs may not use the
same layout for every chunk.

PiperOrigin-RevId: 547297911
Change-Id: If3136327241b509b6c77617123fa38aa2ab0e2ac