Use `void*` instead of `absl::Status*` for ElementwiseFunction
Previously, most functions that accepted an `ElementwiseFunction` used
`ElementwiseFunction<N, absl::Status*>`, accepting an `absl::Status*`
pointer as an extra argument.

In some cases an argument that is not an `absl::Status*` may need to
be passed, however.  While it is possible to still pass any pointer
using `reinterpret_cast`, it is clearer to just use a `void*`
parameter.

PiperOrigin-RevId: 551386488
Change-Id: I9c0ec5f10bbeea3983e4f667f915e28a7f4d6070