Refactor data type operations and add scalar comparison operation
For operations that only deal with the raw bytes, a single
implementation is now instantiated for each size/alignment combination,
rather than for each individual type.

The 2 equality comparison kinds are now specified by the
`EqualityComparisonKind` enum.

This commit also removes support for "same value" comparison, which
was used for zarr v2 fill value comparison.  Zarr v2 fill value
comparison is now done using "identically equal" comparison, which has
a user-visible effect only in the case that a NaN value is specified
as the fill value:
- In that specific case, the exact NaN representation will now be
preserved; in previous versions, if any NaN is specified as the fill
value, and the entire chunk is written with NaN values, then the chunk
is not stored.  When read back, whichever NaN value was specified as
the fill value is returned, which may differ from the actual NaN
values written.

Array <-> scalar comparison operations are now supported in addition
to the array <-> array comparison operations.  These are used by
`AsyncWriteArray` to compare to more efficiently compare to a scalar
fill value, particularly in the `IterationBufferKind::kContiguous`
case where a scalar fill value would force a downgrade to the
`IterationBufferKind::kStrided` representation.

PiperOrigin-RevId: 551386634
Change-Id: Ifddb725a14eb39651d55fca90d4d78764139feab