Add ChooseReadWriteChunkShapes function
This is used for joinly picking read and write chunk shapes for
the zarr v3 sharded format.

Also changes chunk size selection based on aspect ratio to not exceed
the target number of elements if possible; previously, the chunk size
was chosen to minimize the difference between the actual number of
elements and the target number of elements, which could result in the
actual number of elements exceeding the target number of elements.

PiperOrigin-RevId: 553632221
Change-Id: I04785070ca906e3f12dc9fe00091d192b31a84c9