Add `assume_cached_metadata` open mode and fix `assume_metadata` semantics
Previously, `assume_metadata` was designed to serve several different
use cases:

1. Creating a TensorStore from multiple machines concurrently without
   attempting to read/write the metadata separately from each
   machine.  In this case the stored metadata may not actually exist,
   and might also be concurrently created via a separate open call
   using a shared cache.
2. Re-opening a previously-created TensorStore as part of a
   distributed pipeline from many different workers.

To satisfy use case (1), because the concurrent creation via a
separate open call may lead to the metadata cache indicating that the
metadata does not exist, the metadata re-validation logic needs to
ignore negative cache entries.  At the same time, in order to support
use case (2) and behave like a normal TensorStore, operations like
resize need to be supported and new metadata needs to be taken into
account.

Previously, as a compromise, negative metadata cache entries were
handled specially, but this behavior was not entirely consistent.
Additionally, due to a bug, `TensorStore::spec()` did not work when
the TensorStore was opened with `assume_metadata`.

While it was originally intended that `assume_metadata` would be used
automatically when re-opening a TensorStore as part of
deserialization, this would have been problematic due to the
inconsistent handling of the metadata cache.

This commit resolves these issues by splitting `assume_metadata` into
two separate modes:

- `assume_metadata`, which never accesses the stored metadata and does
  not support metadata-mutating operations like resize, and serves use
  case (1);

- `assume_cached_metadata`, which doesn't access the stored metadata
  when opening but otherwise accesses it normally, and serves use
  case (2).

This commit also fixes `TensorStore::spec()`.

In a subsequent commit, `assume_cached_metadata` will be used
automatically for deserialization.

PiperOrigin-RevId: 559971981
Change-Id: Ifa8d3fc257c23701e1eae9f04f08e7fbbcd68edd