Add support for weak references to cache entries
A WeakPinnedCacheEntry reference prevents the entry from being
destroyed due to memory pressure but does not prevent the entry's
parent cache from being destroyed if all strong references are
released.

This can be held as a member of another cache entry to ensure that one
cache entry outlives another, without resulting in a memory leak as
would occur if a strong reference (`PinnedCacheEntry`) were used in
that way.

PiperOrigin-RevId: 561233994
Change-Id: I0e1ea81043fbf19a3df051dfa337fdfbc6d78219