Fix cache entry insertion race condition
Previously, the cache pool mutex could incorrectly be released after
checking that a cache entry is not present, but before adding the new
entry.

Consequently, it was possible for there to exist simultaneously more
than one cache entry for the same key, with the entry inserted second
being leaked.

This fixes a memory leak detected a small fraction of the time by
`tensorstore/kvstore/ocdbt:zarr_test`.

PiperOrigin-RevId: 561234419
Change-Id: Id379f3f1741e911780d033eb824974cc15b1cc89