Add IndexTransformGridPartition::GetCellTransform method
The existing `PartitionIndexTransformOverGrid` function iterates over
both the `grid_cell_indices` and corresponding `cell_transform`.

The `GetGridCellRanges` function, used for storage statistics
computation, provides only ranges of `grid_cell_indices`, but not the
corresponding cell transforms, to avoid wastefully computing it even
when not needed.

The newly added `GetCellTransform` method will be used in a susequent
commit that adds zarr v3 support to obtain the cell transform
afterwards in cases where it is needed.

PiperOrigin-RevId: 564262593
Change-Id: I3e774d0be9653ed55e43ecd4e866699bf1014c5f