Fix bugs in kvs_backed_cache / kvstore transaction mechanism
- Previously, if within a single transaction, the same key was written
by one ChunkCache entry and then read (but not modified) by another
ChunkCache entry, then write would be lost.

- Previously, if one cache entry performed a transactional read that
was handled by a writeback request to a prior cache entry for the same
key in the same transaction, and an `if_not_equal` condition was
specified and not satisfied, a `set_cancel` callback that was supposed
to be unreachable would incorrectly be invoked, leading to a crash.

PiperOrigin-RevId: 595784166
Change-Id: Ie060614623c9541a31934e037bc757dab5e53695