Change ElementwiseFunction to use 2-d buffers in place of 1-d buffers
This improves efficiency when an array cannot be decomposed into long
1d buffers but can be decomposed into relatively large 2d buffers.

This also changes ElementwiseFunction to just return `bool` to
indicate success/failure rather than an `Index` to indicate the number
of elements successfully processed.  The partial success information
was not actually needed, and added significant complexity, especially
with 2-d buffers.

PiperOrigin-RevId: 595784752
Change-Id: I61fe9d04869581954112e7cb898a3173255de169