Fix bug in OCDBT numbered manifest cache
Previously, when writing a new numbered manifest, the future
corresponding to the update operation could complete before the
updated manifest was stored in the cache entry, leading to the stale
previous manifest being read by the subsequent `ReadVersion` call used
to check that the write was successful.  This in turn leads to an
error due to the version that was expected to have been written not
being present.

This error occurred consistently when running ocdbt:driver_test
OcdbtBasicFunctionalityTest.ReadWriteOps.{"manifest_kind":"numbered"}
with --runs_per_test=10000.

PiperOrigin-RevId: 595796295
Change-Id: Iff120d804ee36ad9cf0b9c2435ce52c3ab4524dc