Use template parameter pack for tensorstore::{Read,...} options
Previously, the options for
`tensorstore::{Read,Write,Copy,ResolveBounds,Resize}` were specified
as a single struct.

With this change, in addition to the struct the options can also be
specified as a template parameter pack, and are identified by the
type.  This matches how options are specified to `tensorstore::Open`
and `tensorstore::kvstore::Open` and similar APIs.

To enable identifying the options by their type, a wrapper struct is
now used for the progress functions.

PiperOrigin-RevId: 599337804
Change-Id: Ic6ec6d3afe8c837cf81c27f2af69035e2796a9ca