zarr v2: Eliminate unnecessary copies when decoding chunks
Previously, when decoding a compressed chunk, the entire chunk was first
decompressed into a temporary cord, and then an additional copy was made
into the destination array.

An unnecessary copy was also made when decoding an uncompressed chunk
from a non-flat cord.

With this change, these unnecessary copies are avoided except for a
multi-field array, which are expected to be rare.

PiperOrigin-RevId: 617711170
Change-Id: Id8521c053e0e4b4d655f011af2a4a8364935050d