Fix NumPy array conversion writable constraint
Previously, when converting from a NumPy array, the opposite constraint
on the writable flag was used, incorrectly leading to copies when not
required, and not making copies when required.  The latter case, of
failing to make a copy when required, such that an array that was
intended to be immutable was modified, is unlikely to have occurred
in practice since it could only be triggered by writing to a TensorStore
object created by calling `ts.array(x)`, where `x` is a NumPy array not
marked writable.

PiperOrigin-RevId: 640773299
Change-Id: Ied033259e253bbe4907cd7df1c7028842ffb3b6d