Compute open kvstore cache keys from spec prior to opening
This also avoids the extra work of even opening the KvStore driver if it is
already present in the open kvstore driver cache.

This fixes an issue with the OCDBT driver where the spec (and
consequently the cache key) would include additional configuration
information from the cache when available, such that multiple open calls
with identical specs could result in separate driver instances, leading
to inefficient writes. In the particular case where the underlying
storage does not support atomic writes, this could also lead to race
conditions when writing.

PiperOrigin-RevId: 642735685
Change-Id: I2803954cead41211029603db6c91d8effdecc625