Prevent spurious ABORTED errors with OCDBT numbered manifest
In the case of contention when writing numbered manifests, the OCDBT
driver would fail with an error of ABORTED rather than retrying.

This also fixes a separate bug, in which the wrong set of manifests are
deleted when the number of manifests exceeds
kNumNumberedManifestsToKeep (128).  This could have led to data loss in
the case of contention when writing numbered manifests, but due to the
other bug all such cases would lead to an ABORTED error rather than
silent data loss.

Additionally, this factors out the kvstore concurrent write tests into a
single common implementation that is now also tested on OCDBT.

PiperOrigin-RevId: 644618647
Change-Id: I44986f5839b4c2cc53fb4e18a814c97f2ad5d8ec