Refactor AsyncWriteArray to support chunk-varying fill value
Previously, AsyncWriteArray::Spec stored a per-chunk fill value, which
made it incompatible with rectilinear grids where chunk shapes are not
uniform.  Furthermore, in principle this meant that the fill value was
"tiled".  No drivers actually supported "tiled" fill values, and instead
the fill value was always uniform over all chunked dimensions.

With this change, `AsyncWriteArray::Spec` no longer stores information
about the chunk shape, and instead the shape is passed to all
`AsyncWriteArray` methods that require it.  In a future change, this
will be required in order to support rectilinear chunk grids.

With this change, the fill value is stored as a `SharedOffsetArray` with
a domain that must contain the domain of every individual chunk. This
means that tiling of the fill value is no longer supported, but a fill
value that varies across chunks is supported. For example, in a 3-d
array of shape `[height, width, 4]`, where the last dimension
corresponds to RGBA, it might be desirable to make the fill value `[0,
0, 0, 1]` (broadcast to the shape of the array). Previously, this would
only be possible if the last dimension is not chunked. With this change,
it is supported regardless of whether the last dimension is chunked.

PiperOrigin-RevId: 645465010
Change-Id: I694b061a56976c114242022506fce625299b7cbc