Fix bazel pytest samefile issue and remove unnecessary --pyargs option
On Windows, in some cases pytest collection fails due to
temporary local-spawn-runner files created and deleted by concurrent
test runs.  This fixes that by monkey patching `os.path.samefile` in the
bazel pytest runner.

Additionally, the bazel pytest integration always passes in filenames,
not module/package names, and therefore the --pyargs option, which
causes pytest to first check if the argument is a module or package
name, is not correct.

PiperOrigin-RevId: 653735048
Change-Id: I05afa937f234f6b8c1f06f3ff2e7031c043b5731