bazel_to_cmake: For proto aspects, add FilesProviders for the generated proto files.
This attempts to fix issues related to https://github.com/google/tensorstore/issues/197

The basic issue appears to be errors in the translation from bazel to CMake
when generating various grpc proto outputs.  This is still a best effort
as bazel_to_cmake doesn't support aspects in the same way as bazel.

PiperOrigin-RevId: 674415224
Change-Id: I1dcdc0ff200c206f5e9e4a13753f0aed331e161f