Introduce context "file_io_locking" to select locking for the "file" driver.
When "mode" is set to "os", (default) flock or equivalent os locking is used to guard writes.

When "mode" is set to "lockfile", files are opened using posix flags
O_CREAT | O_EXCL to ensure single process ownership of a lock file
for a limited duration.

When "mode" is set to "none", no locking is used by the "file" driver, and corruption may occur.

PiperOrigin-RevId: 694224794
Change-Id: I78e53c0ea0e9dc8a9ac36df97cb4792609ba99db