avif_writer: correctly set matrix coeffs w/non-yuv 444
Setting the matrix coefficients to identity is only valid for YUV 444.
Previously YUV 400 was supported, but this would create an invalid
bitstream (that still may decode with libaom, and dav1d without
strict_std_compliance set).

A more strict check was introduced the libaom encoder to prevent
creating invalid bitstreams:
e107085626 ensure subsampling is 0 w/AOM_CICP_MC_IDENTITY

From the 1.0.0 with Errata 1 version of the spec [1]:

[1]: 6.4.2. Color config semantics If matrix_coefficients is equal to
MC_IDENTITY, it is a requirement of bitstream conformance that
subsampling_x is equal to 0 and subsampling_y is equal to 0.

PiperOrigin-RevId: 730991804
Change-Id: I8a9cce15cc801f375a456448a51b5cb6379cc4cd