Fix tests' dependencies on gtest implementation details.
These tests rely on the exact failure messages generated by `ElementsAre()`, which are not part of gtest's public API and are subject to change without notice. If a test needs to verify matcher messages, it must only use matchers owned by the same project. It's never OK to depend on implementation details of code that one doesn't own.

PiperOrigin-RevId: 738045085
Change-Id: I5f84775808f751fa41aa76609394424649e027f1