Add DriverKindRegistry and improve unregistered kvstore error message
Previously, if the user attempted to create a KvStore spec using a
TensorStore driver id, a generic "unregistered id" error was returned.

With this change, the error message will indicate that a
TensorStore id rather than a KvStore id was specified.

In a subsequent commit, the reverse case (creating a TensorStore spec
from a KvStore driver id) will handled (not as an error) via the new
format auto-detection driver.

PiperOrigin-RevId: 755690932
Change-Id: Ia712f089a45aee2931ef01f19046f94426c02544