kvs_backed_chunk_driver: Fix transactional open
Previously, when performing a transactional open, the metadata was
correctly retrieved from the transactional node of the metadata cache,
but only a non-transactional read request for the metadata was issued
prior to checking the transactional cache entry.  In common cases the
metadata is already in the transactional metadata cache node, but in
certain cases it may not be, incorrectly leading to the metadata not
being found.

With this change, the logic of requesting the optionally-transactional
metadata is factored out and used for both `ResolveBounds` and `Open`.

PiperOrigin-RevId: 761356472
Change-Id: I31c2a37b8e1d8bee93c6ea7f617ac3a7121b9926