Fix stack driver serialization issue
Previously, the serialization (used for pickling) implementation for
an open TensorStore driver attempted to apply the
`assume_cached_metadata` option by first querying the existing open
mode of the spec and then adding `assume_cached_metadata`.  However,
for the stack driver, the open mode is returned as `OpenMode{}` if not
all layers have the same open mode, which leads to deserialization
attempting to apply an invalid open mode.

To fix this, the open mode option handling now treats open modes that
contain neither `open` nor `create` as "additive"; this allows
`assume_cached_metadata` to be added without needing to query the
existing open mode.

PiperOrigin-RevId: 818983954
Change-Id: Ifcd99b9939ae243d9687455f2b90133d209d73c7