Convert batch Request type from tuple<> to a struct type with traits detection.
While working on other batch requests I found that I wanted to try a ReadRequest type with two string fields. Access to the fields was by type in a few places, and since ReadRequest was implemented as a tuple<...>, duplicate field types are problematic, so I've removed that aspect of ReadRequest.

PiperOrigin-RevId: 832726371
Change-Id: Id6ce5c31edb4f7e16fc79f47bd4a1d2e3534dda7